--- 08_other_function/src/020_practice_portals/start/Example.jsx
@@ -1,33 +1,0 @@
-import { useState } from "react";
-import Toast from "./components/Toast";
-
-const Example = () => {
-  const [toastOpen, setToastOpen] = useState(false);
-
-  return (
-    <div>
-      <h3>
-        トーストの作成（createPortal）
-      </h3>
-      <p>ボタンを押すと.container.start要素にマウントされて、表示されるトーストを作成してください。トーストにはToastコンポーネントを使用してください。</p>
-
-      <div className="container start"></div>
-      
-      <button
-        type="button"
-        onClick={() => setToastOpen(true)}
-        disabled={toastOpen}
-      >
-        トーストを表示する
-      </button>
-      {toastOpen && (
-          <Toast
-            visible={toastOpen}
-            handleCloseClick={() => setToastOpen(false)}
-          />
-      )}
-    </div>
-  );
-};
-
-export default Example;
