+++ 08_other_function/src/020_practice_portals/start/components/Toast.tsx
@@ -1,6 +1,11 @@
 import "./Toast.css";
 
-const Toast = ({ visible, handleCloseClick }) => {
+type ToastPropsType = {
+  visible :boolean
+  handleCloseClick : (value: React.MouseEvent<HTMLButtonElement>) => void
+}
+
+const Toast = ({ visible, handleCloseClick }: ToastPropsType) => {
   const toastClassName = visible ? "toast is-visible" : "toast";
   return (
     <div className={toastClassName}>
